+++ jda/src/jvmMain/kotlin/dev/schlaubi/lavakord/jda/Utils.kt
@@ -5,5 +5,5 @@ import dev.schlaubi.lavakord.LavaKord
 /**
  * Calculates the shard id for the guild with [snowflake] as its id.
  */
-public fun LavaKord.getShardIdForGuild(snowflake: Long): Int =
-    ((snowflake shr 22) % shardsTotal).toInt()
+public fun LavaKord.getShardIdForGuild(snowflake: ULong): Int =
+    ((snowflake shr 22) % shardsTotal.toUInt()).toInt()
